backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Pillow==11.0.0
python-dotenv==1.0.0
requests==2.31.0